# api/docs/swagger.yaml
definitions:
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      error: {}
      message:
        type: string
      requestId:
        type: string
      status:
        type: string
    type: object
...
  models.Configuration:
    properties:
      description:
        type: string
      field_type:
        type: string
      filter:
        type: string
      name:
        type: string
      value:
        type: string
    type: object
...
  models.SignInRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.SignInResponse:
    properties:
      access_token:
        type: string
      expired_at:
        type: string
    type: object
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://kassa360.uz
paths:
  /client-api/auth/sign-in:
    post:
      consumes:
      - application/json
      description: Sign In
      operationId: sign_in
      parameters:
      - default: '"Asia/Tashkent"'
        description: Location
        in: header
        name: location
        type: string
      - description: Sign In
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SignInResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Sign In
      tags:
      - Auth
  /client-api/configuration/default_values:
    get:
      consumes:
      - application/json
      description: Default Values Configuration
      operationId: default_values_configuration
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Configuration'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Default Values Configuration
      tags:
      - Configuration
  /client-api/configuration/field_types:
    get:
      consumes:
      - application/json
      description: Field Type Configuration
      operationId: field_type_configuration
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Configuration'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Field Type Configuration
      tags:
      - Configuration
  /client-api/configuration/group_types:
    get:
      consumes:
      - application/json
      description: Group Type Configuration
      operationId: group_type_configuration
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Configuration'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Group Type Configuration
      tags:
      - Configuration
  /client-api/configuration/regex:
    get:
      consumes:
      - application/json
      description: Regex Configuration
      operationId: regex_configuration
      parameters:
      - description: Regex
        in: query
        name: regex
        required: true
        type: string
      - description: Value
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Configuration'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Regex Configuration
      tags:
      - Configuration
  /client-api/configuration/validation_functions:
    get:
      consumes:
      - application/json
      description: Validation Function Configuration
      operationId: validation_function_configuration
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Configuration'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Validation Function Configuration
      tags:
      - Configuration
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"